+++ modules/pubmatic/openwrap/database/database.go
@@ -8,7 +8,6 @@ import (
 type Database interface {
 	GetAdunitConfig(profileID, displayVersion int) (*adunitconfig.AdUnitConfig, error)
 	GetActivePartnerConfigurations(pubID, profileID, displayVersion int) (map[int]map[string]string, error)
-	GetPubmaticSlotMappings(pubID int) (map[string]models.SlotMapping, error)
 	GetPublisherSlotNameHash(pubID int) (map[string]string, error)
 	GetWrapperSlotMappings(partnerConfigMap map[int]map[string]string, profileID, displayVersion int) (map[int][]models.SlotMapping, error)
 	GetPublisherVASTTags(pubID int) (models.PublisherVASTTags, error)
