+++ modules/pubmatic/openwrap/database/mysql/slot_mapping_test.go
@@ -11,156 +11,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func Test_mySqlDB_GetPubmaticSlotMappings(t *testing.T) {
-	type fields struct {
-		cfg config.Database
-	}
-	type args struct {
-		pubID int
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    map[string]models.SlotMapping
-		wantErr bool
-		setup   func() *sql.DB
-	}{
-		{
-			name:    "empty query in config file",
-			want:    map[string]models.SlotMapping{},
-			wantErr: true,
-			setup: func() *sql.DB {
-				db, _, err := sqlmock.New()
-				if err != nil {
-					t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-				}
-				return db
-			},
-		},
-		{
-			name: "empty site_id in pubmatic slotmappings",
-			fields: fields{
-				cfg: config.Database{
-					Queries: config.Queries{
-						GetPMSlotToMappings: "^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)",
-					},
-				},
-			},
-			args: args{
-				pubID: 5890,
-			},
-			want: map[string]models.SlotMapping{
-				"adunit": {
-					PartnerId:    1,
-					AdapterId:    1,
-					VersionId:    0,
-					SlotName:     "adunit",
-					MappingJson:  "{\"adtag\":\"0\",\"site\":\"0\",\"floor\":\"0.00\",\"gaid\":\"0\"}",
-					SlotMappings: map[string]interface{}{"adtag": "0", "floor": "0.00", "gaid": "0", "owSlotName": "adunit", "site": "0"}, Hash: "", OrderID: 0,
-				},
-			},
-			wantErr: false,
-			setup: func() *sql.DB {
-				db, mock, err := sqlmock.New()
-				if err != nil {
-					t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-				}
-				rows := sqlmock.NewRows([]string{"slot_name", "pm_size", "pm_site_id", "ad_tag_id", "ga_id", "floor"}).
-					AddRow("adunit", "300x250", "", 234, 3, 0.12)
-				mock.ExpectQuery(regexp.QuoteMeta("^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)")).WithArgs(5890, models.MAX_SLOT_COUNT).WillReturnRows(rows)
-
-				return db
-			},
-		},
-		{
-			name: "duplicate slotname in pubmatic slotmappings",
-			fields: fields{
-				cfg: config.Database{
-					Queries: config.Queries{
-						GetPMSlotToMappings: "^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)",
-					},
-				},
-			},
-			args: args{
-				pubID: 5890,
-			},
-			want: map[string]models.SlotMapping{
-				"adunit": {
-					PartnerId:    1,
-					AdapterId:    1,
-					VersionId:    0,
-					SlotName:     "adunit",
-					MappingJson:  "{\"adtag\":\"111\",\"site\":\"555\",\"floor\":\"0.51\",\"gaid\":\"5\"}",
-					SlotMappings: map[string]interface{}{"adtag": "111", "floor": "0.51", "gaid": "5", "owSlotName": "adunit", "site": "555"}, Hash: "", OrderID: 0,
-				},
-			},
-			wantErr: false,
-			setup: func() *sql.DB {
-				db, mock, err := sqlmock.New()
-				if err != nil {
-					t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-				}
-				rows := sqlmock.NewRows([]string{"slot_name", "pm_size", "pm_site_id", "ad_tag_id", "ga_id", "floor"}).
-					AddRow("adunit", "300x250", 123, 234, 3, 0.12).
-					AddRow("adunit", "400x350", 555, 111, 5, 0.51)
-				mock.ExpectQuery(regexp.QuoteMeta("^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)")).WithArgs(5890, models.MAX_SLOT_COUNT).WillReturnRows(rows)
-
-				return db
-			},
-		},
-		{
-			name: "valid pubmatic slotmappings",
-			fields: fields{
-				cfg: config.Database{
-					Queries: config.Queries{
-						GetPMSlotToMappings: "^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)",
-					},
-				},
-			},
-			args: args{
-				pubID: 5890,
-			},
-			want: map[string]models.SlotMapping{
-				"adunit": {
-					PartnerId:    1,
-					AdapterId:    1,
-					VersionId:    0,
-					SlotName:     "adunit",
-					MappingJson:  "{\"adtag\":\"234\",\"site\":\"123\",\"floor\":\"0.12\",\"gaid\":\"3\"}",
-					SlotMappings: map[string]interface{}{"adtag": "234", "floor": "0.12", "gaid": "3", "owSlotName": "adunit", "site": "123"}, Hash: "", OrderID: 0,
-				},
-			},
-			wantErr: false,
-			setup: func() *sql.DB {
-				db, mock, err := sqlmock.New()
-				if err != nil {
-					t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-				}
-				rows := sqlmock.NewRows([]string{"slot_name", "pm_size", "pm_site_id", "ad_tag_id", "ga_id", "floor"}).
-					AddRow("adunit", "300x250", 123, 234, 3, 0.12)
-				mock.ExpectQuery(regexp.QuoteMeta("^SELECT (.+) FROM giym.publisher_slot_to_tag_mapping (.+)")).WithArgs(5890, models.MAX_SLOT_COUNT).WillReturnRows(rows)
-
-				return db
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			db := &mySqlDB{
-				conn: tt.setup(),
-				cfg:  tt.fields.cfg,
-			}
-			got, err := db.GetPubmaticSlotMappings(tt.args.pubID)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("mySqlDB.GetPubmaticSlotMappings() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
 func Test_mySqlDB_GetPublisherSlotNameHash(t *testing.T) {
 	type fields struct {
 		cfg config.Database
