+++ modules/pubmatic/openwrap/metrics/config/metrics_test.go
@@ -156,7 +156,7 @@ func TestRecordFunctionForMultiMetricsEngine(t *testing.T) {
 	adFormat := "banner"
 	dealId := "pubdeal"
 	host := "sv3:xyz1234"
-	getTime, sendTime, requestTime := 300*time.Millisecond, 300*time.Millisecond, 300*time.Millisecond
+	getTime, sendTime := 300*time.Millisecond, 300*time.Millisecond
 	queryType := models.AdunitConfigQuery
 
 	// set the expectations
@@ -205,7 +205,6 @@ func TestRecordFunctionForMultiMetricsEngine(t *testing.T) {
 	mockEngine.EXPECT().RecordVideoImpDisabledViaConnTypeStats(publisher, profile)
 	mockEngine.EXPECT().RecordGetProfileDataTime(endpoint, profile, getTime)
 	mockEngine.EXPECT().RecordSendLoggerDataTime(endpoint, profile, sendTime)
-	mockEngine.EXPECT().RecordRequestTime(endpoint, requestTime)
 	mockEngine.EXPECT().RecordDBQueryFailure(queryType, publisher, profile)
 	mockEngine.EXPECT().Shutdown()
 
@@ -258,7 +257,6 @@ func TestRecordFunctionForMultiMetricsEngine(t *testing.T) {
 	multiMetricEngine.RecordVideoImpDisabledViaConnTypeStats(publisher, profile)
 	multiMetricEngine.RecordGetProfileDataTime(endpoint, profile, getTime)
 	multiMetricEngine.RecordSendLoggerDataTime(endpoint, profile, sendTime)
-	multiMetricEngine.RecordRequestTime(endpoint, requestTime)
 	multiMetricEngine.RecordDBQueryFailure(queryType, publisher, profile)
 	multiMetricEngine.Shutdown()
 }
