+++ modules/pubmatic/openwrap/metrics/mock/mock.go
@@ -501,18 +501,6 @@ func (mr *MockMetricsEngineMockRecorder) RecordRequestAdPodGeneratedImpressionsC
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RecordRequestAdPodGeneratedImpressionsCount", reflect.TypeOf((*MockMetricsEngine)(nil).RecordRequestAdPodGeneratedImpressionsCount), arg0, arg1)
 }
 
-// RecordRequestTime mocks base method
-func (m *MockMetricsEngine) RecordRequestTime(arg0 string, arg1 time.Duration) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "RecordRequestTime", arg0, arg1)
-}
-
-// RecordRequestTime indicates an expected call of RecordRequestTime
-func (mr *MockMetricsEngineMockRecorder) RecordRequestTime(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RecordRequestTime", reflect.TypeOf((*MockMetricsEngine)(nil).RecordRequestTime), arg0, arg1)
-}
-
 // RecordSSTimeoutRequests mocks base method
 func (m *MockMetricsEngine) RecordSSTimeoutRequests(arg0, arg1 string) {
 	m.ctrl.T.Helper()
